Added property page, it is still not finished. I do not know if I would
Added property page, it is still not finished.  I do not know
if I would like to have the properties applied to all of the windows
or the properties applied only to the terminal where the property
page was invoked.

Another option is a checkbutton: "Global apply changes" in the
property page.  Ideas on this issue are welcome.

Miguel.

svn path=/trunk/; revision=716