Fixed the /tmp security issue as posted by Robert Wilhelm to the gnome-list
This is still broken in the sense that temp files are created with mkstemp()
which are then never removed. Has to do with hacked interface to the
gtk-xmhtml widget for handling images grabbed via http. No one uses
the help-browser as a web client anyways, so at least its secure if someone
foolishly chose to use it as such.

Dr Mike <msf@redhat.com>

svn path=/trunk/; revision=1268